Rename generate_*.py to plural form
