+++ eiffel-syntax-and-usage/event-schemas.md
@@ -120,4 +120,4 @@ _examples:
 ```
 
 ## Generating schemas and documentation
-Schema and documentation files should never by updated by hand. Instead, modify the schema definition file and use the [generate_doc.py](../generate_doc.py) and [generate_schema.py](../generate_schema.py) Python scripts to regenerate all schemas and documentation. The included [makefile](../Makefile) calls the scripts with the required arguments. Just make sure you've installed the Python dependencies specified in [requirements.txt](../requirements.txt) before running the scripts. The schema definition files and the generated files should always be in sync, i.e. changes should always be committed together.
+Schema and documentation files should never by updated by hand. Instead, modify the schema definition file and use the [generate_docs.py](../generate_docs.py) and [generate_schemas.py](../generate_schemas.py) Python scripts to regenerate all schemas and documentation. The included [makefile](../Makefile) calls the scripts with the required arguments. Just make sure you've installed the Python dependencies specified in [requirements.txt](../requirements.txt) before running the scripts. The schema definition files and the generated files should always be in sync, i.e. changes should always be committed together.
